pyproject.toml
[project]
name = "codelens"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]

[project.scripts]
codelens = "codelens.main:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

